add/trash_loc,userdb생성,sqlite3적용, common앱에 models.py 추가
