+++ templates/trash_loc_list.html
@@ -6,7 +6,7 @@
     <title>Trash Locations</title>
 </head>
 <body>
-    <h1>Trash Locations</h1>
+    <h1>쓰레기통목록</h1>
 
     <table border="1">
         <thead>
@@ -35,7 +35,7 @@
                 </tr>
             {% empty %}
                 <tr>
-                    <td colspan="8">No data available</td>
+                    <td colspan="8">데이터가 없습니다</td>
                 </tr>
             {% endfor %}
         </tbody>
